Fix workout CRUD bugs and add March 10 workout
- Fix parseInt for workout ID when starting new workout
- Fix getRelativeDate() strict equality with (int) cast
- Add today's gym session: 16 sets, 10,630 lbs volume

+++ templates/history.php
@@ -13,7 +13,7 @@ function formatDuration($started, $ended) {
 
 function getRelativeDate($timestamp) {
     $now = time() * 1000;
-    $diff = floor(($now - $timestamp) / 86400000);
+    $diff = (int) floor(($now - $timestamp) / 86400000);
     
     if ($diff === 0) return 'Today';
     if ($diff === 1) return 'Yesterday';
